fixes for cs
